backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
deepdiff==6.7.1
PyYAML==6.0.1
pydantic==2.4.2
